Try going all in on hc, start seeing cracks
