+++ src/lib/api/index.ts
@@ -1,0 +1,1 @@
+export { apiClient } from './api.ts';
